fix(api): removed hoc wrappers
HOCs made more sense when we wanted this library to act as wrappers for other components. The
philosophy of the library has changed since then, and HOCs don't make sense any longer. They were
causing bugs through the use of PureComponent + context
