+++ src/index.js
@@ -2,9 +2,7 @@
 import * as srcUtils from './utils'
 
 export { default as Grid } from './grid'
-export { default as GridHOC } from './grid.hoc'
 export { default as Layout } from './layout'
-export { default as LayoutHOC } from './layout.hoc'
 export { version } from '../version'
 export const utils = srcUtils
